Support added for SeeMeCNC HE280 extruder.
SeeMeCNC Rostock Max V2/V3 Delta Printer with HE280 Hotend uses an accelerometer integrated into the end effector platform for the bed leveling. SeeMeCNC only supported this to Rpetier 0.92.2. This branch ports the HE280 support to Repetier 1.0.4.

Note that the included configuration is for a slightly modified Rostock Max. It was a V2 upgraded using the SeeMeCNC V2->V3 kit, with the HE280 replacing the original PEEK hotend. I have also added a second extruder, two filament sensors, and two pairs of wires to drive the hotend fan and cooling fan directly. The z-probe is connected to the Z-MIN endstop pin.  This is further described in https://forum.repetier.com/discussion/8324/rostock-max-he280-accelerometer-z-probe-filament-runout-sensors